requirements.txt
starlette~=0.41.3
PyJWT~=2.10.1
aiomysql~=0.2.0
bcrypt~=4.2.1
Jinja2~=3.1.4
colorama~=0.4.6
uvicorn
python-multipart
itsdangerous
nest-asyncio~=1.6.0
email-validator
pydantic~=2.11.7